Add header_files Directory
This commit will add the functions.h and variables.h header files which
are dependencies of the investment.c file being the program entry point
file. Some struct functions with struct function pointer parameter
variables were used in the functions.h so as to be able to store data
globally in the global struct datatype definition or instantiation of
the global struct data-type declaration. All global variables were
expressed in the variable.h header file.